Remove bad lines
